Splitting into proper components and helpers
